+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 3.4.3 (unreleased)
 ------------------
 
+- Update faceted xml configuration. Add a workflow "draft" filter on projects. (Don't keep projects in draft state)
+  [boulch]
+- Add a project_directly_submitted checkbox field in ideabox controlpanel to choice is project can be directly submitted
+  [boulch]
 - Add campaign_emails field on campaign type to send mails only on to these campaign managers
   [boulch]